Updated mutex w/ new sync api.

+++ us/include/twz/mutex.h
@@ -3,7 +3,7 @@
 #include <twz/sys.h>
 
 struct mutex {
-	_Atomic int sleep;
+	_Atomic long sleep;
 };
 
 #define MUTEX_INIT                                                                                 \
